+++ Proyect-TitanLife/src/main/java/aiss/model/repository/UserRepository.java
@@ -46,7 +46,7 @@ public class UserRepository {
         return u;
     }
 
-    public User addUser(String username, String email, String password, String retype, Integer age, Integer height, Integer weight, Integer hip, Integer waist) {
+    public User addUser(String username, String email, String password, String retype, Integer age, Double height, Double weight, Double hip, Double waist) {
         User u = new User(username,email,password,retype,age,height,weight,hip,waist);
         users.put(username,u);
         log.log(Level.INFO, "El usuario " + u.getUsername() + " fue creado");
